requirements.txt
streamlit
notebook
tensorflow
keras
torch
torchvision
openai
langchain
langchain-openai
langchain-community
python-dotenv
pyPDF2
flask
pillow
scikit-learn
numpy
matplotlib
seaborn